Updates to the HTML and CSS files
HTML: I reordered the buttons in terms of semantics by adding them to an unordered list <ul>.

Regarding the CSS file, I had to adjust the button class so it affects the <li> elements by adding .a.button. Also, there where modifications to the margins and the addition to a list-container class to delete the bullet points of the unoredered list.